Refactoring 'children' attribute in yaml file
Children can be both 'dependencies', i.e. instantiated components or actual delegated 'children' which have to be created by the component. Splitting this argument into two makes it a lot clearer to understand what is going on. This involves changes to all yaml files, many drivers, odemisd and the model.