gui comp NavigableBarPlotCanvas & NavigablePlotViewport: reorganise to reduce numbers of refresh
Currently when there is a new data, the data is passed to the plot, which has a hack to not immediately display it.
Then adjusted X and Y ranges are computed, both of which trigger a redisplay of the data.
In case no range are adjusted, it could happen that the data is never displayed.

So now, we directly pass to the canvas all the values/ranges needed to properly display the plot.
Also, by using subcsribers instead of setters for the VAs, we avoid triggering a redisplay if the ranges
are not changed (as the subscribers are not called on setting the same value again).

Also, ensure that when windowPeriod is present on a stream, it's used to compute the "fitting" range in X.

This seems to fix the data range for chronograms sometimes only updated on the second data.