gui comp canvas PlotCanvas: simplify by always computing the data range
The current code would delay computing the data range to when the plot will
be displayed. However, in most cases, the data is eventually displayed
(it's only for some corner cases where the data would be updated so fast that
it's updated again before the next display). Moreover, in many cases, the
ranges are already known, so there don't need to be computed again.

A lot of code was there just to handle this, while it's much simpler to
immediately compute the ranges, and then expect that the data is present.
Moreover, data_width and data_height was computed too, while it's actually
just a subtraction. => just compute it when needed.

Also don't check every time that the data is ordered. If the caller passes
wrong data, weird things will happen. That makes set_data() quite a bit faster.

Also simplify the set_ranges() function to trust the inputs are not None.

This should have no behaviour change.