driver semcomedi: restart acquisition thread if it failed
If the acquisition thread failed, it exists. That's very rare, but we have once encontered
an (unexplained) assert failure, which caused that. In such case, for now, the driver needs
to be restarted, so the entire Odemis must be restarted.

Now, on the next acquisition, a new thread will be created.

Also improve the logging for the assertions, so that hopefully next time, it'll be
clearer what's happening.