odemisd: increase timeout during instantiation
Some hardware (eg, some Andorcam2) can take a looong time to initialise.
In particular, when there are multiple Andorcam2, it can take more than 2 min to initialise.

It'd be pretty hard to break the init() code into sub calls, so for, now just increase the
timeout to 3+ min.
However, for the standard functions calls, we now restrict the timeout to 30s. In theory,
more than 5s should be enough for any call... but in practice, there could be some long call
in some drivers, so let's play it safe.