gui cont tabs Sparc2Align: move fibaligner.y if FAV_POS_ACTIVE is available
The new fibaligner also supports referencing on the Y axis.
Automatically referencing it is easy, but that means we need to store the
"good" position somewhere. => use FAV_POS_ACTIVE.
We mostly do exactly the same as for the X position, but that position
has to be recorded on the spectrometer-selector, so it's duplicated.