gui util img: simplify val_to_pos* functions
Same as the simplification in commit 86b560d750 (... simplify by always computing the data range)

No need to pass the data_height/data_width around.

At the same time, we fix a small bug when the spectrum contained only 0 data,
which confused the test for data_heigh is None, but it also picked when it was 0.