gui acquisition: Ensure that presets info is always updated in the main GUI thread
Whenever a VA from a hardware handled by a stream changes, the presets is updated.
This typically happens when the user changes them (so, in the main GUI thread).
However, this can happen in other contexts, in particular during the acquisition,
when the stream or the OPM changes some settings. In that case it would happen in
a separate stream. In some rare cases, this could crash/freeze the GUI.
As only the combox box is updated, and usually it's updated to the _current_ value,
it'd trigger an error only very rarely.