driver semcomedi: handle case when scale * res is (slightly) bigger than shape
The .scale and .resolution setters ensure that scale * res is always <= shape.
However, due to floating point error, it happens that even if shape/scale <= res,
that previous state is not true. For example:
shape = 5760, res = 1025, scale = 5.619512195121952
int(shape/scale) * scale == 5760.000000000001

Actually, 5760/1025 ~= 5.619512195121951, but due to rounding error, the client
passed 5.619512195121952, and this triggered ratio > 1.
=> detect such a case, and just "reduce" the ratio to 1.