acq align autofocus: Sparc2AutoFocus sorts the detectors
For most of the functionality the order of the detectors doesn't matter.
However, for detecting the light, not all detectors are as good at it.
So we use alphabetical order, to allow forcing which detector will be used,
from the YAML file.
It also helps to make the call reproducible.