gui cont tabs: Don't update mirror FAV_POS_ACTIVE if it's obviously wrong
This is actually a hack of a hack. Normally, the user can only
access the alignment tab if the mirror is engaged, so it should never
get close from the parked position. However, for maintenance, it's
possible to hack the GUI and enable the tab even if the mirror is
not engaged. In such case, if by mistake the mirror moves, we should
not set the "random" position as the engaged position.

Especially important now that the metadata is persitent.