driver andorcam3: parameter to restrict the maximum binning
Allow to limit the maximum binning.
Also introduce BASELINE_COR, to change the baseline metadata if the hardware is improperly configured.

This is to workaround a bug in some Zyla cameras which tend to subtract too much the values
when doing full vertical binning. So the binning would be limited to a small value (eg, 16),
and the baseline reduced (a little bit). This way, the binning will be automatically done
in software by the CompositedSpectrometer, and not subtracted too much. With respect to the SNR,
there shouldn't be any difference, as the CMOS don't gain anything from hardware binning (but bandwidth).