driver spectrometer: Fix baseline subtraction
There were two issues with the baseline subtraction:
 * It multiplied the baseline by the number of column instead of lines (plain bug)
 * If the total baseline was too big, it would adjust the data so that the minimum is
   at the expected baseline level, while updating the metadata as if everything was fine.

This second behaviour caused weird jumps of the data depending on the minimum value
(which is typically noise). So it's better to just subtract the baseline, but ensure
that the minimum value is >= 0. In case the total baseline is too big, the official
baseline level is adjusted in the metadata too.