Bug fix temporal spectrum data assembly
In a recent pr the data assembly on the temporal spectrum data was changed, as it was incorrect.
The data shape was x, y, 1, which was introduced by new code regarding the long exposure time, which were on the same branch.
However, this was done at a incorrect location in the code. If we want to keep the new data assembly (which is correct), we need to add a third dimension to the data.
Code is tested now on HEAD.