driver omicronxx: also disable ad-hoc messages on the master
The sending of ad-hoc messages seems to only be governed by the "master"
device. However, that was the only one were we didn't disable it.
So it had no effect. In most cases, it's not a big issue, as the ad-hoc
messages are automatically discarded next time a response is read.
However, if many ad-hoc messages have been sent, they fill up the queue
of the serial port (in the Linux kernel), and prevent to get the
responce corresponding to the latest command.