gui conf: force minimum wavelength step to 1nm
Add support to specific key_step_min for the _axes_, in the exact same way as it was done for the VAs.
Use it to specify that the minimum wavelength step should be 1nm (instead of 1pm).
It's more convenient on small wavelengths, but even more importantly, that ensures that the
default prefix for wavelength 0 is "n", instead of "p".