gui plugin: log when the plugin is taking control
In theory, the plugins can modify any part of the GUI, and so can have effect anywhere.
In practice, most of the plugins use the convenience functions to be triggered when the
user selects a specific menu entry, and to show an acquisition dialog.

As we don't control the code of the plugins, we cannot ensure that the logging messages
give enough information. In case of bug, that's a pity. However, it's at least possible
to log every time the plugin function is called, when it's done from a standard entry point.