gui comp viewport: Ensure that the spectrum and chronogram legends are always shown if there is data
The change to a VA for the .hrange and .vrange (in commit e450ca1c1c50) helped reduce
the number of times the legend is updated. However, when the data was cleared, the legend
would be hidden, and if the same data was shown again, it wouldn't detect that the legend
needed to be displayed again.

=> Explicitly set .hrange and .vrange to None every time the data is changed to ensure that
the next time some data is shown, the legend range is updated again, allowing it to be shown.