acq stream sync: Explicitly set MD_DIMS of SEMSpectrumMDStream spectrum data
Until now, MD_DIMS wasn't set. That would mean only the shape would be relied on to guess the dimensions.
However, if the detector set the MD_DIMS, which is now the case for the standard spectrometer,
which set (properly) the dims to "XC", that metadata is copied, and it turns out to be incorrect
on an array of C11YX.

In practice, this prevented the display of the spectrum data after an acquisition.

=> Explicitly set MD_DIMS to CTZYX.