driver actuator MultiplexActuator: fix very first position update not reported
The initial position was not using the _applyInversion(). So .position VA contained exactly
the ._position object. That meant that on the next update, the VA would already contain the new
value and so think that it's not worthy notifying.
On the following calls, as _applyInversion() was used, the object would be different, and this
wouldn't have the issue.

=> _applyInversion() at init too.

This fixes issues where the metadata would sometimes be incorrect if the hardware has just been moved once.
For instance, this could happen for the filter position of the SPARC with MD_OUT_WL.