gui conf util: do not create SettingEntry if it shouldn't be shown
If the control type defined in the configuration is CONTROL_NONE, then
no widget is created for the given VA. In such case,
create_setting_entry() used to still create a SettingEntry, with no
widget. The goal was to allow storing/restoring the VA before/after
acquisition.

However, in all the cases so such hidden VA, so far, they shouldn't
be shown because they don't matter for the user (and often they are
read-only). So it doesn't make sense to store/restore them. It causes
extra memory usage.

Also discard read-only VAs in get_global_settings_entries(), instead of
inside the presets. It confused the acquisition window to think
that the settings changed every time one of these settings. That's
a little annoying as it includes the ccd.temperature (which is
frequently refreshed).