Bug fix streakcam: bug in MD_TIME_RANGE_TO_DELAY handling
Choices defined in yaml file for timeRange are different from actual choices on HW VA due to floating point issues.
When setting a new value for the delay, the MD was updated by updating the value for the corresponding timeRange.
However, due to the floating point issue, the timeRange from the HW VA was slightly different from the values in the MD, which are based on the yaml file.
Now check for the closest and then for almost equal to overwrite the value, but keep the same key. Warn if new key is added and check that length of MD fits with length of choices on HW VA.