acq stitching: weaver bugfix
In python2, numpy.divide returns an int (places after dot are cut off) and in python3 a float. When using the old code in python3, it gives a weight matrix of slightly different shape and the following operations fail. To avoid this, numpy.arange with the case distinction is replaced by numpy.linspace, which always gives the right shape.