gui conf: show CCD EM gain on the SPARC
It seems that for some CCDs (eg, Andor Newton), the EM CCD gains are independently set from
the readout rate and pre-amp gain. In such case, it's handy for the user to control it from the GUI too.

On the SECOM, so far, all the systems with an EM CCD have a LUT, and can automatically set the EM gain.