gui comp viewport: fix spectrum scale locks sometimes activating automatically
The locks should _only_ happen when the user explicitly zooms/move the scale.
However, the current code was taking a shortcut, and in some cases, it would
consider the range changed was due to the user while it was caused by other code
adjusting the range. In particular, when the data contains only 0 values on a side,
the scale is automatically adjusted to fit the data. This would trigger the lock.
Also, when quickly changing between spectra, it would sometimes update the VA before the data
is updated, which would cause some kind of race-condition and lock the scale too.

=> Change to force the scale lock from every GUI event, corresponding to the user
interaction. This way, it can only be trigger on user action.