plugins timelapse: store data in parallel, to save time
When the requested acquisition rate is close from the acquisition time of a frame,
it helps to store the data in parallel (while already acquiring the next frame).

Typically, the time to store a single 2D image is less than 0.1s, so the improvement is only really
noticeable when the period is < 1s.