acq path: warn if there is no pass-through filter on the SPARC
On the SPARC, there should always be a "pass-through" filter position (ie, no filter at all).
It would work without this position, but it'd much harder for the user to do the alignment,
as the signal would be very limited.
Normally, all hardware have this position, but there can be typo in the microscope file
such as a missing affect, or the position being incorrectly named, which would prevent the
OPM from properly selecting it.