plugins timelapse: optimised acquisition if the period is small and only one stream is used
In case the user actually wants to acquire a series of images as fast as possible, the
standard code was not adapted. It had a lot of overhead.
One example is when the user wants to do some kind of super-resolution acquisition on FM.

=> detect such case and hijack the pretend that the Stream plays live, but instead of
showing the images, we save them. It's quite ugly, but that allows to support a wide
range of streams easily, and with quite a short code.