acq: automatically unblank the beam during calibration too
Also make the hack even more hacky by only activating it on the SECOM.
On the SPARCs with blanker support, this would work "fine" for the SEM streams,
but as the MDStreams are not aware of the blanker, they would always leave it active.

Eventually, it's clear that this approach is not really right, and it's just a hack.
It should be changed to have the CompositedScanner to take care of automatically turning
on/off the blanker.