gui evt: Fix Knob Events with wxPython 4
We had a "super clever" way to generate events.
So clever that the new "super clever" functions of wxPython 4 got completely confused,
and the attributes were lost.

=> Make everything explicit. It makes it work on wxPython 4, and it's also more obvious
so that you don't have to be "super clever" at the time that you read the code in order to understand it.

It should still be compatible with wxPython 3 (but untested)