gui cont settings: Fix failure to show acquisition date on foreign locale
If the system's settings are set to show dates in a non-ascii locale, the date
failed to be shown. That's because on Python 2 strftime always returns bytes (string).
Later on, the add_readonly_field() tried to convert it to unicode, but assuming "ascii".

=> Convert to unicode using the current encoding (eg, UTF8)

To reproduce, one could use such command:
LC_ALL=zh_CN.UTF-8 odemis-gui --standalone