driver actuator LinearActuator: don't move at init if already referenced
If the axis was already referenced, it would try to move to the "current" position,
but that was without including the offset. It would cause an actual move, unneeded,
and sometimes even prevent initialisation, if the current position of the child is outside the
range.

=> Don't move if there is no need to move.

Also clean-up a little the log messages.