gui cont tabs: don't fail if no MD_PIXEL_SIZE or MD_POS
Let's not entirely fail from showing the data.

Probably, instead of using .get(MD_, hard-coded value), the best would be
to automatically fill the metadata in the static stream at init, and then all the
rest of the code can expect to find the metadata, and so use the same fallback value.