integration tests: bugfix python interpreter
Passing the interpreter as an argument doesn't work. If the script is run in python3 and we start odemis with python2 (and vice versa), we get a pyro error. Maybe the pythonpath gets confused in this situation. Instead, we can just start the scripts with python2 or python3 and use sys.executable to start odemis.
The integration tests also need to be stored should be stored in separate folders for python2 and python3, however it's fine to use a single file for the summary.