plugins cli_rgb: attempt to immediately stop the acquisition when cancelling
When cancelling, instead of stopping the progress bar but leaving the acquisition on, now
we explicitly try to stop the current part of the acquisition. If the
user is trying to cancel because the dwell time was too long, it's
really useful to not let an acquisition run the background, and messup
all the future acquisitions.