Distance tool
A new functionality is added in the GUI which allows to measure the distance between 2 points on an image. It's possible to delete it by pressing the "Delete" key.
When exporting to "print-ready", the rulers are visible.

Also refactor the Label code, to make it more independent (and usable by the Ruler class).