gui comp RulerOverlay: Do not get scared if the shift is a list instead of a tuple
Sometimes the shift is a (Notifying) list, instead of being a tuple.
We could argue that the canvas should only have tuple as position (ie, immutable),
but anyway, let's make the code handle also list in addition to tuple.
Especially, here, this is only used to check whether the canvas has moved/rescaled compared
to the last time we drew.