gui comp RulerOverlay: Fix moving the ruler after editing it using Ctrl+drag
v_start_pos and v_end_pos should be Vec, not simple tuples, as they afterwards
we use the sum operator (+) to modify it.
It's actually a little bit of a trouble with this Vec class, because tuple already support the + operator,
for concatenation. So doing tuple + Vec "works", but gives a tuple of 4 elements instead of doing a sum.

It would (later) cause errors like this:
08:53:21 ERROR   main           : Traceback (most recent call last):
  File "/home/piel/development/odemis/src/odemis/gui/comp/overlay/world.py", line 1370, in on_motion
    self._selected_ruler.on_motion(vpos, ctrl_down)
  File "/home/piel/development/odemis/src/odemis/gui/comp/overlay/world.py", line 950, in on_motion
    self._view_to_phys()
  File "/home/piel/development/odemis/src/odemis/gui/comp/overlay/world.py", line 908, in _view_to_phys
    self._calc_edges()
  File "/home/piel/development/odemis/src/odemis/gui/comp/overlay/world.py", line 1012, in _calc_edges
    ex, ey = self.v_end_pos
ValueError: too many values to unpack