gui util img test: instantiate a wx.App() also for the TestARExport
It appears to have been broken by commit c620e5d9760b (Distance tool),
but that's probably just because the change to the test cases caused
a reorder of the run. Now TestARExport is run early, before any other
test create a wx.App(). This is necessary as some of the export functions
use wxPython calls (which require an App to be available).

=> create explicitely a wx.App() at class start-up.
Also remove a few of these creations where it is not necessary anymore, as
the GuiTestCase already create one.