bugreporter: add a timeout to the external commands
When using just the future.result() timeout, on timeout the command output
is not provided at all, and the command is not stopped.
So now, we explicitly stop the command after the timeout, which allows to get the
output back.
Especially useful for odemis-hw-status.txt which can sometimes block on one component,
but the rest is still interesting.