driver hamamatsurx: don't log if no data is received
Sometimes the connection goes wrong, and the recv() constantly immediately returns,
with no data.
We shouldn't log that there is no data, and immediately retry, as it has a tendency to
fill up the log.
=> handle this case explicitly.

We might want to look further at this issue, and do something more clever in this case,
but we first would need to understand what triggers this.