driver tmcm: fix to the reference switch initialisation
It hasn't ever been properly tested. If a switch was on, it would actually
fail because some attributes are missing so early. In addition, it
would not even look for the correct switches (mixed up between axis number and
switch number).

=> The simulator now reports such a switch being turned on, to force the code
to be run during testing.
=> Move the code later in the init.

Also remove old TODO already implemented.