gui comp: also fix stream.add_readonly_field()
Commit e07417ad8 fixed setings.add_readonly_field(), but the code is
duplicated in the stream controller => also fix it there.

Also simplify the code, because unicode() can pretty much never fail
on random object... and if it fails it's not useful to have a message
that uses str(obj).

Also clarify a little the comment, as str is actually unicode, it's a
little confusing.