acq stream StaticARStream: warn if a combination of pos+polarization is missing
Normally for every polarization, we acquire every positions. However,
there are two reason this doesn't happen:
 * Something went wrong the acquisition file
 * Some positions are slightly different (due to floating point error),
   and not detected as similar.

The stream doesn't really mind, but if a non-existing combination is
requested, it's obviously going to fail. So let's try to warn it early,
and if that due to the second case, we might eventually try to make a
more clever code.