acq stream test: fix exposure time setting
Now that there is the special integrationTime VA on the SPARC streams,
we have to be more careful on how the exposure time is set.
When exposureTime is not passed as detvas, integrationTime is
automatically created and used. So the raw detector .exposureTime is
*never* used, as we used to assume previously.
This caused many test cases to use the default exposureTime of 1s,
instead of way smaller values. So the test cases were very long to run.

Also merge/fix some of the live update test cases.
Also fix some of the integrationTime test cases.

Note: there is still a bug when live update is used at the same time at
integrationTime (on the metadata)