acq stream sync: fix live update and integrated images
When several acquisitions are used to build a single integrated image,
the live update would pass to the helper stream only the final
integrated image.

This is wrong for 2 reasons:
 * The helper stream would try to integrate each received image, showing
   extra integration.
 * The live image would be updated only very unfrequently

=> update the live image after each (sub-) acquisition.
Note: to avoid some kind of "blurring" effect between pixel, we also explicitly reset raw[]
before the beginning of a series of integration. All the streams use
the same format of .raw, so this should be safe.