gui util wxadpater: Add another workaround for StaticText
The issue with StaticText also appears when changing the content while
the widget is hidden. We probably could workaround it by changing all
the code that do this, and force the widget to be shown, but it feels
safer and more correct to just ensure that .Show() fully show the
widget.