acq path: fix the check for pass-through filter
We tried to check for the presence of the pass-through filter, but a
typo prevented from ever finding it (choices -> choices.values()).
Anyway, the function needs to also account for the possible different
role of the ccd (eg, ccd0).

So instead of having a dedicated function, just detect the case *in* the
code that is in charge of the move to the right filter. This avoids
having to code twice the same thing in two different ways.