gui comp canvas: Also update the display if the canvas is disabled
It was an optimisation because typically if the canvas is disabled,
the user doesn't need to see the images... however that's not true.
For instance, in the SPARC we are going to get live update during
acquisition. During the acquisition the canvas is disabled, but we
still want to see the image.
=> Don't optimise, and always show images when the streams provide
new data.