driver hamamatsurx: support for transp
Allows to swap/invert axes, which is quite important for the
time/wavelength to be correct.

Also add metadata MD_DIMS=TC so that when displaying the raw data Odemis
shows it in a proper way (eg, when doing a snapshot).

Tested on the real hardware and seems to work fine.