gui cont tabs SPARC2AlignTab: when leaving the tab, always ensure the calibration light is off
The calibration light is turned during focusing.
It would normally be turned off automatically when cancelling the
autofocus. However, on the streak cam, if the manual focus was active,
the light would not be turned off.

=> When leaving the tab, reset the manual focus (so that when coming back to the tab
it doesn't look like we are in this mode), and turn off the light.