driver: Add missing b's
We didn't get enough honey, so adding extra b's.

The migration to Python 3 missed a couple of places where the string
should now be bytes.
That happened mostly in the serial port opening functions because these
are different in the simulator and the actual hardware.

Several of them caused the driver to block forever when connecting to a
(real) device.