driver hamamatsurx: Migrate to using bytes for Python 3
In Python 3, the network sockets send and receive bytes.

Also adjust a couple of extra places in the message handling.
In particular, the CameraInfo now behaves a little bit less special, and
each line is return as a separate return value instead of one value
containing a list.

These changes haven't be tested on the hardware yet.