config: don't force utils to run with Python 3 by default
There are two reasons we don't want them to run with Python 3:
 * The code has been more tested with Python 2, so like the rest of
   Odemis, it should use Python 2 by default for now.
 * For the scripts which try to connect to the backend, it will fail if
   the backend is running in a different Python version, so better use
   Python 2, which is the default for Odemis.

We still ship "adjusted" script (ie, with python3 shebang) in a separate
directory for running explicitly in Python 3.
It should also be possible to just type for example: python3 /usr/bin/tmcmconfig

To implement it, it's a little complicated, because there are 2 layers
that try to update the "shebang": "python setup.py", and dh_python3.