model vattribute: also accepts numpy integers in IntVA and ResolutionVA
Numpy integers easly creep everywhere. Normally that's fine because they
quack like an int (except for the over- under-flows of course, but that
typically doesn't happen in the number we handle).
So let's accept them also in the VAs.

Note: in Python 2, numpy.int* were a subclass of int, so it worked
fine (not for numpy.uint* though). In Python 3, this "link" has been removed,
that's why we have to explicitly add it