acq stream ARRawProjection: fix cache when data has polarization
The projection caches the azimuthal projections, so that supseccend
display of the same position (and polarization) are fast.
However, when not finding a given position/polarization, it would reset
completely the cache for the position. If there were already
polarizations computed, they would all be gone.

=> Only reset/create a new dict for a position if it doesn't exist yet.

So now, looking back & forth between two polarizations at the same
position in the GUI is faster.
The drawback is that it will increase memory usage.