acq stream test StaticARStream: fix polarization test on Python 3 and clean-up
Some of the polarization test cases failed on Python 3. That's because
the test cases would compare two polarizations "at random". However,
this randomness is just based on the set, and on Python 2, it's every time
the same polarizations which come. On Python 3, this is much less
predictable. The current fake data was not very different, and for some
polarization the projection to RGB lead to the same (black) image.

=> Improve the data generation to be really different for every
polarization.

Also factorize the generation, so that all the test cases use the same
code to get fake AR data.

Also improve the waiting code, so that it cannot block forever in case
of issue.