driver avantes: improve acquisition event handling
Based on a proper FSM, with all the events that can happen in the same
queue. More code, but it's much clearer that it should work in all
cases, without race conditions.

One specific improvement is that the acquisition is stopped early if
unsubscribed before receiving the data.

Also add extra test cases to test most of the common code paths.