bugreporter: fix Python 3 str vs bytes issues
csv handler uses *str*, so no need to try to convert to bytes.

On the contracty, Request() needs bytes... at least for the data (but
not for headers).

Also, when opening a file in "a+" mode, in Python 3 the position is set
at the end, so reading the file brought nothing.