acq stream projection: RGBSpacialProjection support tint for RGB raw data
We used to no care about the tint, but instead, it makes more sense to
multiply each channel independently based on the tint.
In particular, if a "RGB" data has all the channels identical (ie, it's
actually greyscale), then the behaviour of tint is the same as for a
greyscale image.