model vattribute: value of ListVA should always be a NotifyingList
When setting the value of a ListVA, there was a short moment during
which the .value was the original object, before being converted
into a NotifyingList. In particular, the subscribers would always get
the raw object. That meant that if you were writting a tuple, the .value
would be sometimes a tuple and sometimes a list. Confusing.

=> immediatly convert to a NotifyingList, so that the object is always
the same.