driver smartact: fix MCS2 initialisation and Exceptions on Python 3
Reading the DLL version and hardware name hadn't been tested, and it
countained bugs.
=> Fix them to use the proper API.

Also fix issues with the Exceptions on Python 3.
=> use the same style as in the other drivers (eg andorcam2, avantes).

If the device is not found, raise a HwError, so that the backend shows
a nice error message and suggests the user to try again.

Also do a few minor clean-ups.