Export of pyramidal picture in higher resolution
When exporting pyramidal images, the current tiles in the projection do not necessarily correspond to the maximum possible resolution (but instead, they just fit the current zoom level in the GUI). Typically, the user wants lots of pixel, when he has a lot of pixels.
=> Try to use the maximum resolution provided by the raw data. At least up to 5000x5000 px.

Also update the unit tests to check for such cases.