acq align delphi: fix stages not found
The change from children to dependencies broke that part of the code.
A long long time ago, the stages where not directly available in the
model. So we had to rely on a hack, which consisted in looking on the
children of the main stage. The recent change to use depedencies broke
this part of the code.
=> Don't rely on this hack at all, and instead use the standard way to
search for components. That's what the manual calibration has been
doing, so it works on all systems, for sure.