model core: No need to warn if we have plenty of margin to listen to VAs
prepare_to_listen_to_more_vas() showed a warning if it's reached the
limit of possible files to open.
In practice, this limit is very high, so this only happens if something
else (eg Eclipse) has already changed the soft limit. No need to show
warnings in such case, as everything is fine.