gui cont Sparc2AlignTab: Handle better cancelation of manual focus
Two (weird) user behaviours were not well handled:
1. The user quickly toggle/untoggle the manual focus button (a couple of
times). In such case, every click would be executed, at the same time.
Eventually the right (last) thing would happen, but for a while
the gauge would get crazy.
=> Cancel the previous manul focus action before starting a new one.

2. The user starts manual focus, and immediately after presses auto
focus. It would result in both running, then the auto focus taking
control.
=> Immediately disable the auto focus button when the user toggles
the manual focus. It's now possible to also immediately change the
grating and detector. That's fine.

Also, when changing the tab while the manual focus (or autofocus) is
running automatically cancels it. However, the progress bar is left
as-is. So when coming back, the progress bar conveys the meaning that
something is still running.
=> Reset the progress bar when coming back to the tab.