acq path: don't queue OPM setPath() calls
Sometimes, for instance when changing quickly between chamber and
alignment tab, setPath() is called several times in a row.
Currently, every request is full-field, which can end up being very
slow. However, we only care about the latest requested position.

=> Discard move requests which haven't started yet.

We could be even more clever, and immediately stop the current moves,
but that's quite a lot more complicated, espcially because that means
all the callers need to handle CancelledError.