gui util img: make AR export image at least 800px
The projection size depends on the input size. On small input size, the
output image ends up really tiny (eg 128px), which makes all the legend
and spider web almost unreadable. In the GUI, it's fine, as the image is
scaled to the view.

=> Force the output image to be at least 800px.

Also move the outer circle to exactly touch the image (instead of being
spaced by a 1.5 tick). Otherwise it gives the feeling that some data is missing.

Also clean-up a little the code to make it a little bit more direct and
explicit.