gui comp miccanvas: fix RGBA DataArray caching
As DataArrays cannot be directly used as keys in a dict, because they
are not hashabel, we used id() to proxy them. Although id() can NOT collide
with other objects currently allocated, it can collide with previous
objects.

When opening a new acquisition file (in the analysis tab) this caused sometime
the image shown to be the one from a previous stream instead of the new
stream, because that new stream create a DataArray at the same space in
memory as the previous stream. This might have caused a lot more issues
with the pyramidal stream because they keep reusing memory space for
tiles... however a bug prevented the code from ever using the cache.

=> Use weakref's. It's less convenient as they cannot be put in a dict,
but it ensures they are never mixed up.

Also clear the cache when calling .clear(), because in this case we
already know we won't need it, and so we free a bit of memory earlier.