gui main: don't try to speed up Pyro connections
With the current behaviour of Pyro, the proxies are not reused in the
rest of the GUI, so the connections opened are immediately closed.
So instead of saving time, this causes extra delay.
When starting with a large backend, it ended up costing 1s extra, over 7s.

=> Remove "speed-up" hack.

Also move the first log to as early as possible to avoid confusing log
from the early start with the end of the previous log.