model core: prepare_to_listen_to_more_vas() increases the file limit by a larger amount
The function was very optimistic, expecting that each VA needs only one
file opening. With a large back-end, the GUI end-up still reaching the
file limit.

In practice, all included, it seems that subscribing to a
VA (which was not proxy'd before) causes about 6 new files to be opened.
This is measured by using this command which reports the amound of files
currently opened by the GUI:
lsof -p $(pgrep -f odemis.gui.main) | wc -l

Comparing between the SettingsObserver activated or not, this showed
that for 102 VAs we need 686 more files.

=> Bump the file limit by 6 for each VA to be opened.