util spectrum: fix get_wavelength_per_pixel() failing on 2D data with MD_WL_POLYNOMIAL
Only very old acquisitions were stored using the MD_WL_POLYNOMIAL, but
we still officially support reading them.
Using this type of metadata, if the wavelength was not on the first
dimension, as is the case of 2D data of dimensions XC, then the
generated range would be incorrect.

Note: on the long term, we should just convert polynomials to MD_WL_LIST
when reading them, and dropping the metadata entirely.