acq stream projection: LineSpectrumProjection clean-up
Clarify the function, especially to look the same as the other
projection functions.
It also changes the returned data to follow the standard behaviour of
having the minimum value at the first index, which was not the case for
the X dimension.

The users will have to be adjusted (but anyway, there are only 2, and
half of them are wrong)