driver phenom: on error during loading, make sure to stop the updater
Without this, as soon an error occurs, the updater keeps running,
which keeps showing more error.

Also, if an error is reported, but it actually moved to the requested
position, we keep going.
This happens sometimes, if for some reason the focus of the navcam is
considered incorrect, then SEM loading would constantly fail.